New Screen Fader Method (Commit #6)
Now there are 2 screen faders. One that simply alters the alpha level using a script and another that uses animations and coroutines. Press T or Y for various effects. Camera shaker is not finished, but it "works". After midterms, I'll get around to finally solving the problem as camera shaker is not urgently needed.