Additional pause implementation and addition of SFX/Music tags.
If you add true to the end of any PlaySound call, the sound it makes will not be paused if the game is paused. Useful for menu SFX and background music.
The SFX and Music tags don't really do anything yet, but they allow scripts to tell if an AudioObject is music or SFX. Could be useful for objects.
This does mean that PlaySound is a 4/5-argument function call with a pre-call so long that it can't fit in Visual Studio's default editor width. I am looking into a solution for that.