Pushing Test Scene (Volume Options)
Here's a test scene for volume options. Note that the sfx for hitting a note was removed so I put it back as it's own sound that's on loop so I could distinguish what was music and what was sfx. There are three sliders, top is for master, middle is for bgm, and bottom is for sfx (I think).