Added FarmVN and FarmRhythm
These will be the scenes to contain the new first farm scene and the Farm Rhythm game. These scenes contain barebones prefabs, but lack all asset implementation at this time. The only work done on this branch should be filling in the missing pieces to make the scenes work. Also contains a DialogueScriptSetter script, but it doesn't do anything right now. Don't use it, but don't delete it either.