Chibis In beat update
Chibis update thier sad faces faster to make it have a stronger feedback quicker. Calculations for sad faces are noe reworked so they update when the score increases as well.

e.g. when missing multiple notes in a row, faces will stay sad instead of flipping on and off. Hitting a note afterward resets the face to the current mood.

Other: Cleaned up some code and added docs for some.