Finished day 2 Bar and Kitchen dialouge files
Day 2 Kitchen should be tested and ready. There's still the bug where Maria doesn't show up, but I have no idea how to fix that.

Day 2 Bar isn't tested, but all the dialouge is in. It has a line that indicates where the scene transition should be, and a lot of exits and fancy macro usage stuff. Big source of bugs right there.